smear-cursor: remove the Lua timer bridge and refresh perf tooling
Finish the host-bridge and timer cleanup, tighten the remaining conceal/timer edge cases, and refresh the perf-reporting scripts and snapshots around the new flow.\n\nThis also captures the workspace and Nix tooling updates that were made alongside the final cleanup.

+++ nvim/rust/plugins/smear-cursor/src/events/handlers/viewport.rs
@@ -129,7 +129,7 @@ pub(super) fn maybe_scroll_shift_for_core_event(
     let viewport_row_shift = match delta.vertical_rows {
         Some((previous_top_row, current_top_row)) => screen_distance(
             window,
-            i64::from(current_surface.window_size().max_row()),
+            current_surface.window_size().max_row(),
             previous_top_row,
             current_top_row,
         )?,
